Split up selected and unselected
